pyproject.toml
[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.poetry]
name = "vizme-backend"
version = "0.1.0"
description = ""
authors = ["pavivin <pavivin@yandex.ru>"]
readme = "README.md"
packages = [{include = "vizme_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
blacksheep = "^1.2.12"
SQLAlchemy = {extras = ["asyncpg"], version = "^2.0.12"}
httpx = "^0.24.0"
celery = {extras = ["redis"], version = "^5.2.7"}
orjson = "^3.8.10"
pydantic = {extras = ["email"], version = "^1.10.7"}
alembic = "^1.10.3"
pytest-asyncio = "^0.21.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
factory-boy = "^3.2.1"
async-factory-boy = "^1.0.0"
psycopg2-binary = "^2.9.6"
pyjwt = "^2.6.0"
punq = "^0.6.2"
cryptography = "^40.0.2"
pytest-factoryboy = "^2.5.1"
phonenumbers = "^8.13.11"
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
